+++ bluepysnap/frame_report.py
@@ -245,8 +245,8 @@ class FrameReport:
 
     @property
     def node_set(self):
-        """Returns the node set for the report."""
-        return self.simulation.node_sets[self.to_libsonata.cells]
+        """Returns the name of the node set for the report."""
+        return self.to_libsonata.cells
 
     @property
     def simulation(self):